Fix creation of passive agent nodes on newer SolarWinds versions
- After updating to SolarWinds 2025.1.1, creation of passive agent nodes was
failing, requiring updating to the included agent version (2025.1.1.2078).
This enabled the agent and node to be created, but the ScheduleListResources
operation would fail with '400 Client Error: ProvideFault failed, check fault
information'. Introducing a while loop with try/except, retries, and sleep
timer works around the issue, without inducing excessive delay for sites
where the issue is not seen. Fixes #13.